requirements.txt
coverage==4.5.1
nose==1.3.7
pycodestyle==2.3.1
pydocstyle==2.1.1
PyYAML==3.12
